ssh: don't try to link against libssh2
Part of the point of this is being able to load libssh2 regardless of whether
it was available upon building, as the resulting shared object might move.

Make USE_SSH mean to try to load it at runtime. We still need to know the
signatures of the functions for our macros to work, so include a copy of the
header as a vendored dependency.